CommandAttributes traits
