+++ src/Discord/Builders/CommandBuilder.php
@@ -24,6 +24,8 @@ use function Discord\poly_strlen;
  */
 class CommandBuilder implements JsonSerializable
 {
+    use CommandAttributes;
+
     /**
      * Type of the command. The type defaults to 1.
      *
@@ -39,11 +41,25 @@ class CommandBuilder implements JsonSerializable
     protected string $name;
 
     /**
+     * Localization dictionary for the name field. Values follow the same restrictions as name.
+     *
+     * @var string[]
+     */
+    protected array $name_localizations;
+
+    /**
      * Description of the command. should be emtpy if the type is not CHAT_INPUT.
      *
      * @var string
      */
     protected string $description = '';
+
+    /**
+     * Localization dictionary for the description field. Values follow the same restrictions as description.
+     *
+     * @var string[]|null
+     */
+    protected array $description_localizations;
 
     /**
      * array with options.
@@ -67,126 +83,6 @@ class CommandBuilder implements JsonSerializable
     public static function new(): self
     {
         return new static();
-    }
-
-    /**
-     * Sets the type of the command.
-     *
-     * @param int $type Type of the command
-     *
-     * @throws \InvalidArgumentException
-     *
-     * @return $this
-     */
-    public function setType(int $type): self
-    {
-        if ($type < 1 || $type > 3) {
-            throw new \InvalidArgumentException('Invalid type provided.');
-        }
-
-        $this->type = $type;
-
-        return $this;
-    }
-
-    /**
-     * Sets the name of the command.
-     *
-     * @param string $description Name of the command. Slash command names are lowercase.
-     *
-     * @throws \LengthException
-     *
-     * @return $this
-     */
-    public function setName(string $name): self
-    {
-        if (poly_strlen($name) > 100) {
-            throw new \LengthException('Command name must be less than or equal to 32 characters.');
-        }
-
-        $this->name = $name;
-
-        return $this;
-    }
-
-    /**
-     * Sets the description of the command.
-     *
-     * @param string $description Description of the command
-     *
-     * @throws \LengthException
-     *
-     * @return $this
-     */
-    public function setDescription(string $description): self
-    {
-        if ($this->type == Command::CHAT_INPUT && poly_strlen($description) > 100) {
-            throw new \LengthException('Command description must be less than or equal to 100 characters.');
-        }
-
-        $this->description = $description;
-
-        return $this;
-    }
-
-    /**
-     * Sets the default permission of the command.
-     *
-     * @param bool $permission Default permission of the command
-     *
-     * @return $this
-     */
-    public function setDefaultPermission(bool $permission): self
-    {
-        $this->default_permission = $permission;
-
-        return $this;
-    }
-
-    /**
-     * Adds an option to the command.
-     *
-     * @param Option $option The option
-     *
-     * @throws \OverflowException
-     *
-     * @return $this
-     */
-    public function addOption(Option $option): self
-    {
-        if (count($this->options) >= 25) {
-            throw new \OverflowException('Command can only have a maximum of 25 options.');
-        }
-
-        $this->options[] = $option;
-
-        return $this;
-    }
-
-    /**
-     * Removes an option from the command.
-     *
-     * @param Option $option Option to remove.
-     *
-     * @return $this
-     */
-    public function removeOption(Option $option): self
-    {
-        if (($idx = array_search($option, $this->option)) !== null) {
-            array_splice($this->options, $idx, 1);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Returns all the options in the command.
-     *
-     * @return Option[]
-     */
-    public function getOptions(): array
-    {
-        return $this->options;
     }
 
     /**
@@ -216,7 +112,9 @@ class CommandBuilder implements JsonSerializable
 
         $arrCommand = [
             'name' => $this->name,
+            'name_localizations' => $this->name_localizations,
             'description' => $this->description,
+            'description_localizations' => $this->name_localizations,
             'type' => $this->type,
             'options' => [],
             'default_permission' => $this->default_permission,
